Move into log_test_util
