+++ rpc/connection_test_util.go
@@ -2,7 +2,6 @@ package rpc
 
 import (
 	"errors"
-	"fmt"
 	"io"
 	"net"
 	"time"
@@ -116,28 +115,6 @@ func (eu testErrorUnwrapper) UnwrapError(arg interface{}) (appError error, dispa
 	}
 	return appError, nil
 }
-
-// TestLogger is an interface for things, like *testing.T, that have a
-// Logf function.
-type TestLogger interface {
-	Logf(format string, args ...interface{})
-}
-
-type testLogOutput struct {
-	t TestLogger
-}
-
-func (t testLogOutput) log(ch string, fmts string, args []interface{}) {
-	fmts = fmt.Sprintf("[%s] %s", ch, fmts)
-	t.t.Logf(fmts, args...)
-}
-
-func (t testLogOutput) Info(fmt string, args ...interface{})                  { t.log("I", fmt, args) }
-func (t testLogOutput) Error(fmt string, args ...interface{})                 { t.log("E", fmt, args) }
-func (t testLogOutput) Debug(fmt string, args ...interface{})                 { t.log("D", fmt, args) }
-func (t testLogOutput) Warning(fmt string, args ...interface{})               { t.log("W", fmt, args) }
-func (t testLogOutput) Profile(fmt string, args ...interface{})               { t.log("P", fmt, args) }
-func (t testLogOutput) CloneWithAddedDepth(depth int) LogOutputWithDepthAdder { return t }
 
 // MakeConnectionForTest returns a Connection object, and a net.Conn
 // object representing the other end of that connection.
